Added LADSPA Plugins
LADSPA is an acronym for Linux Audio Developer's Simple Plugin API. It
is an application programming interface (API) standard for handling
audio filters and audio signal processing effects, licensed under the
GNU Lesser General Public License (LGPL). It was originally designed for
Linux through consensus on the Linux Audio Developers Mailing List, but
works on a variety of other platforms. It is used in many free audio
software projects and there is a wide range of LADSPA plug-ins
available.

LADSPA Plugins is a project to port Steve Harris's LADSPA plugins to
Windows OS as DLLs.